# internal/admin/openapi.yaml
openapi: 3.0.3
info:
  title: Go LLM Router Admin API
  description: |
    Comprehensive admin API for Go LLM Router system management.

    This API provides complete administrative control over:
    - Provider management and monitoring
    - User and API key management with RBAC
    - Model discovery and routing
    - Configuration management
    - System monitoring and analytics
    - Real-time WebSocket updates
    - Authentication and authorization management

    ## Authentication & Authorization

    The API supports multiple authentication methods and comprehensive role-based access control (RBAC):

    ### Authentication Methods
    - **Bearer Token**: `Authorization: Bearer <admin-api-key>`
    - **API Key Header**: `X-API-Key: <admin-api-key>`
    - **Query Parameter**: `api_key=<admin-api-key>`
    - **Session Cookie**: For web interface authentication
    - **JWT Token**: For stateless authentication with expiration

    ### Role-Based Access Control (RBAC)
    Users are assigned roles that determine their permissions:

    **Available Roles:**
    - `super_admin`: Full access to all resources and settings
    - `admin`: Full access to most resources, except system-critical settings
    - `operator`: Read/write access to providers and models, no user management
    - `viewer`: Read-only access to all resources
    - `support`: Access to logs, monitoring, and diagnostic tools

    ### Two-Factor Authentication (TFA)
...
    - **Backup Codes**: Recovery codes for TFA
...

...
servers:
...
paths:
...
  /v1/users:
    get:
      tags: [Users]
      summary: List all users
      description: Retrieve a paginated list of all users
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
        - name: q
          in: query
          description: Search query for username or email
          schema:
            type: string
        - name: role
          in: query
          description: Filter by user role
          schema:
            type: string
            enum: [super_admin, admin, operator, viewer, support]
        - name: status
          in: query
          description: Filter by user status
          schema:
            type: string
            enum: [active, inactive, locked, pending]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      tags: [Users]
      summary: Create new user
      description: Create a new user account with specified role
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  temporary_password:
                    type: string
                    description: Temporary password for first login
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user details
      description: Retrieve detailed information about a specific user
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Users]
      summary: Update user
      description: Update user information and settings
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Users]
      summary: Delete user
      description: Delete a user account (cannot be undone)
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /v1/providers:
    get:
      tags: [Providers]
      summary: List all providers
      description: Retrieve a paginated list of all configured providers
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
        - name: q
          in: query
          description: Search query for provider name or description
          schema:
            type: string
        - name: status
          in: query
          description: Filter by provider status
          schema:
            type: string
            enum: [active, inactive, error, unknown]
        - name: sort_by
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, created_at, updated_at, status]
        - name: sort_desc
          in: query
          description: Sort in descending order
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProvidersResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      tags: [Providers]
      summary: Create new provider
      description: Add a new provider configuration
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderRequest'
      responses:
        '201':
          description: Provider created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  provider:
                    $ref: '#/components/schemas/Provider'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /v1/models:
    get:
      tags: [Models]
      summary: List all models
      description: Retrieve a paginated list of all available models
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
        - name: provider
          in: query
          description: Filter by provider
          schema:
            type: string
        - name: q
          in: query
          description: Search query for model name
          schema:
            type: string
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedModelsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # User Management (API Keys) - Updated for RBAC
...
  /v1/users/api-keys/{id}/usage:
    get:
      tags: [Users, API Keys]
      summary: Get API key usage
      description: Retrieve usage statistics for a specific API key
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          description: Time period for statistics
          schema:
            type: string
            enum: [today, week, month, year, all]
            default: month
      responses:
        '200':
          description: API key usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  usage:
                    $ref: '#/components/schemas/APIKeyUsage'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session cookie for web interface

  # Schemas
...